// src/consts/enums/ids/item-ids.enum.ts
export enum ItemIds {
  slimeResidue = 0,
  crabMeat = 1,
  feather = 2,
  turtleShell = 3,
  slimeGoldenCrown = 4,
  knife = 5,
  turtleShellHelmet = 6,
  turtleShellChest = 7,
  turtleShellLegs = 8,
  turtleShellBoots = 9,
  turtleShellGloves = 10,
  stick = 11,
  makeshiftClub = 12,
  joshsHeirloom = 13,
  ratTail = 14,
  skillPointBook = 15,
  apple = 16,
  fishMeat = 17,
  captainsLetter = 18,
  cheese = 19,
  deerPelt = 20,
  wolfFangs = 21,
  vialOfWater = 22,
  ratCatcher = 23,
  woodenBow = 24,
  trophyNecklace = 25,
  machete = 26,
  stoneArrow = 27,
  stone = 28,
}