// src/consts/enums/ids/player-stat-ids.ts
export enum PlayerStatIds {
  mana = 'mana',
  maxMana = 'maxMana',
  attackPower = 'attackPower',
  attackSpeed = 'attackSpeed',
  critChance = 'critChance',
  critMulti = 'critMulti',
  experience = 'experience',
  goldCoins = 'goldCoins',
  level = 'level',
  unspentSkillPoints = 'unspentSkillPoints',
  goldCoinsMultiplier = 'goldCoinsMultiplier',
  manaRegenRate = 'manaRegenRate',
  currentManaRegenTimer = 'currentManaRegenTimer',
  magicDamage = 'magicDamage',
  cooldownReduction = 'cooldownReduction',
  xpMultiplier = 'xpMultiplier',
  shopRefreshCooldown = 'shopRefreshCooldown',
  currentShopRefreshCooldown = 'currentShopRefreshCooldown',
  arrowRecoveryChance = 'arrowRecoveryChance',
  extraFireDamage = 'extraFireDamage',
  extraAirDamage = 'extraAirDamage',
  extraWaterDamage = 'extraWaterDamage',
  extraEarthDamage = 'extraEarthDamage',
  extraLightDamage = 'extraLightDamage',
  extraDarkDamage = 'extraDarkDamage',
  extraPhysicalDamage = 'extraPhysicalDamage',
  increasedSpellDuration = 'increasedSpellDuration',
}